requirements.txt
APScheduler==3.6.3
CacheControl==0.12.10
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
Deprecated==1.2.13
firebase==3.0.1
firebase-admin==5.2.0
gcloud==0.18.3
google-api-core==2.7.1
google-api-python-client==2.40.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-cloud-core==2.2.3
google-cloud-firestore==2.4.0
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.55.0
grpcio==1.44.0
grpcio-status==1.44.0
httplib2==0.20.4
idna==3.3
jwcrypto==1.0
msgpack==1.0.3
oauth2client==4.1.3
proto-plus==1.20.3
protobuf==3.20.0rc1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.14.1
PyJWT==2.3.0
pyparsing==3.0.7
python-dotenv==0.19.2
python-jwt==3.3.2
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
requests-toolbelt==0.9.1
rsa==4.8
six==1.16.0
sseclient-py==1.7.2
tornado==6.1
tzdata==2021.5
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.8
wrapt==1.14.0
